[feat] #6 파일 업로드 크기 제한 로직 수정
